// client/web/src/schema/site.schema.d.ts
export interface VersionContext {
  /**
   * Name of the version context, it must be unique.
   */
  name: string;
  /**
   * List of repositories of the version context
   */
  revisions: VersionContextRevision[];
  /**
   * Description of the version context
   */
  description?: string;
}
/**
 * Description of the chosen repository and revision
 */
export interface VersionContextRevision {
  /**
   * Repository name
   */
  repo: string;
  /**
   * Branch, tag, or commit hash. "HEAD" or "" can be used for the default branch.
   */
  rev: string;
}